Refactored stories.
